* Updated the main menu look to be both fancier and to have better readability (dark text on bright background is better than bright text on dark background, according to research).
* Begun work to remove as many #includes from header files as possible, in favor of forward declarations.

* Removed AntTweakBar again because I managed to fix CEGUI editor. Everything is much easier to work with now, hence I stay with CEGUI.

+++ Engine/Skysphere.cpp
@@ -165,7 +165,7 @@ bool Skysphere::LoadModel(char* filename)
 		fin >> model[i].texcoords.x >> model[i].texcoords.y;
 		fin >> model[i].normal.x	>> model[i].normal.y	>> model[i].normal.z;
 
-		 model[i].texcoords.z = faceID;
+		 model[i].texcoords.z = static_cast<float>(faceID);
 
 		counter++;
 	}
